+++ src/helpers/contract.ts
@@ -2,8 +2,15 @@
 // SPDX-License-Identifier: GPL-3.0-only
 
 import { BlueprintPromise, CodePromise } from '@polkadot/api-contract';
-import { transformUserInput, isNumber } from '../util';
-import type { ApiPromise, BlueprintOptions, InstantiateData, SubmittableExtrinsic } from 'types';
+import { isValidAddress, isValidCodeHash, isNumber } from './util';
+import { transformUserInput } from './callOptions';
+import {
+  ApiPromise,
+  CodeBundleDocument,
+  BlueprintOptions,
+  InstantiateData,
+  SubmittableExtrinsic,
+} from 'types';
 
 export function createInstantiateTx(
   api: ApiPromise,
@@ -43,4 +50,25 @@ export function createInstantiateTx(
   } else {
     throw new Error('Error creating instantiate tx');
   }
+}
+
+export async function getContractInfo(api: ApiPromise, address: string) {
+  if (isValidAddress(address)) {
+    return (await api.query.contracts.contractInfoOf(address)).unwrapOr(null);
+  }
+}
+
+export async function checkOnChainCode(api: ApiPromise, codeHash: string): Promise<boolean> {
+  return isValidCodeHash(codeHash)
+    ? (await api.query.contracts.codeStorage(codeHash)).isSome
+    : false;
+}
+
+export async function filterOnChainCode(api: ApiPromise, items: CodeBundleDocument[]) {
+  const codes: CodeBundleDocument[] = [];
+  for (const item of items) {
+    const isOnChain = await checkOnChainCode(api, item.codeHash);
+    isOnChain && codes.push(item);
+  }
+  return codes;
 }
