+++ src/helpers/index.ts
@@ -1,6 +1,10 @@
 // Copyright 2022 @paritytech/contracts-ui authors & contributors
 // SPDX-License-Identifier: GPL-3.0-only
 
-export * from './instantiate';
+export * from './blockTime';
+export * from './bn';
+export * from './initValue';
+export * from './callOptions';
 export * from './util';
+export * from './contract';
 export * from './chainProps';
