--- src/helpers/instantiate/index.ts
@@ -1,4 +1,0 @@
-// Copyright 2022 @paritytech/contracts-ui authors & contributors
-// SPDX-License-Identifier: GPL-3.0-only
-
-export * from './createInstantiateTx';
