--- src/helpers/util/index.ts
@@ -1,8 +1,0 @@
-// Copyright 2022 @paritytech/contracts-ui authors & contributors
-// SPDX-License-Identifier: GPL-3.0-only
-
-export * from './blockTime';
-export * from './bn';
-export * from './initValue';
-export * from './callOptions';
-export * from './util';
